Here's a Firebase backup-copy
If the firebase database gets messed up, click the options menu to the right of https://gifto-13107.firebaseio.com/ -> Import JSON -> import this file. It will replace the current database with this file. Useful if you add a million entries to the database by mistake, and it stops letting you edit real-time